File update

+++ Encode/Modulator.py
@@ -1,7 +1,5 @@
-import torch
-
 # BPSK Modulator and Add Noise After Modulator
 def bpsk_modulator(codeword):
     codeword = 2 * codeword - 1
 
-    return codeword.to(torch.float)
+    return codeword
